requirements.txt
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
pymilvus>=2.3.0
transformers>=4.30.0
torch>=2.0.0
modelscope>=1.10.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
